refactor: move files in requests/components subdir
