+++ web-client/src/modules/requests/containers/ArchivedRequestsContainer/ArchivedRequestsContainer.tsx
@@ -14,9 +14,9 @@ import {
   InformationModal,
   makeLocalStorageKey,
 } from '../../../../components/Modals/OneTimeModal';
-import Header from '../../components/Header/Header';
-import RequestItem from '../../components/RequestItem/RequestItem';
-import RequestList from '../../components/RequestList/RequestList';
+import Header from '../../components/Header';
+import RequestItem from '../../components/RequestItem';
+import RequestList from '../../components/RequestList';
 
 const ClosedRequestsContainer: React.FC = () => {
   const { t } = useTranslation();
