+++ web-client/src/modules/requests/containers/NewRequestsContainer/NewRequestsContainer.tsx
@@ -22,11 +22,11 @@ import AuthenticationModal from 'src/pages/modals/AuthenticationModal';
 import { AppState } from 'src/store';
 import styled from 'styled-components';
 
-import NewRequest from '../../components/NewRequest/NewRequest';
-import RequestConfirmation from '../../components/NewRequest/RequestConfirmation';
+import NewRequest from '../../components/NewRequest';
+import RequestConfirmation from '../../components/RequestConfirmation';
 import RequestReview, {
   RequestInput,
-} from '../../components/NewRequest/RequestReview';
+} from '../../components/RequestReview';
 import { OpenRequestsLocation } from '../../pages/routes/OpenRequestsRoute/constants';
 
 /* TODO:  integrate with translation if safe */
